notebooks/system_files/rtsp_input/requirements.txt
opencv-python
boto3
python-dotenv
